made code modular by creating react hooks and components in send modal screen
